+++ src/devices/innr.ts
@@ -180,6 +180,13 @@ const definitions: Definition[] = [
         ota: ota.zigbeeOTA,
     },
     {
+        zigbeeModel: ['RB 272 T'],
+        model: 'RB 272 T',
+        vendor: 'Innr',
+        description: 'Smart bulb tunable white E27',
+        extend: [light({colorTemp: {range: [153, 555]}, turnsOffAtBrightness1: true})],
+    },
+    {
         zigbeeModel: ['RB 278 T'],
         model: 'RB 278 T',
         vendor: 'Innr',
@@ -363,6 +370,14 @@ const definitions: Definition[] = [
     {
         zigbeeModel: ['RB 245'],
         model: 'RB 245',
+        vendor: 'Innr',
+        description: 'E14 candle',
+        extend: [light({turnsOffAtBrightness1: true})],
+        ota: ota.zigbeeOTA,
+    },
+    {
+        zigbeeModel: ['RB 243'],
+        model: 'RB 243',
         vendor: 'Innr',
         description: 'E14 candle',
         extend: [light({turnsOffAtBrightness1: true})],