+++ src/devices/nous.ts
@@ -22,6 +22,7 @@ const definitions: Definition[] = [
             {modelID: 'TS0601', manufacturerName: '_TZE200_c7emyjom'},
             {modelID: 'TS0601', manufacturerName: '_TZE200_locansqn'},
             {modelID: 'TS0601', manufacturerName: '_TZE200_qrztc3ev'},
+            {modelID: 'TS0601', manufacturerName: '_TZE200_snloy4rw'},
             {modelID: 'TS0601', manufacturerName: '_TZE200_eanjj2pa'}],
         model: 'SZ-T04',
         vendor: 'Nous',