+++ src/lib/modernExtend.ts
@@ -1557,7 +1557,7 @@ export function quirkAddEndpointCluster(args: QuirkAddEndpointClusterArgs): Mode
 
         inputClusters?.forEach((cluster: number | string) => {
             const clusterID = isString(cluster) ?
-                Zcl.Utils.getCluster(cluster, device.manufacturerID).ID :
+                Zcl.Utils.getCluster(cluster, device.manufacturerID, device.customClusters).ID :
                 cluster;
 
             if (!endpoint.inputClusters.includes(clusterID)) {
@@ -1568,7 +1568,7 @@ export function quirkAddEndpointCluster(args: QuirkAddEndpointClusterArgs): Mode
 
         outputClusters?.forEach((cluster: number | string) => {
             const clusterID = isString(cluster) ?
-                Zcl.Utils.getCluster(cluster, device.manufacturerID).ID :
+                Zcl.Utils.getCluster(cluster, device.manufacturerID, device.customClusters).ID :
                 cluster;
 
             if (!endpoint.outputClusters.includes(clusterID)) {